OEL-101: SearchApiEuropaSearchBackend viewSettings, getSupportedFeatures and draft facets.
